move data repo to :core:data
