+++ rss-generator/core/data/src/main/resources/metadata.yml
